+++ lib/guard/rack/runner.rb
@@ -1,36 +1,55 @@
 require 'fileutils'
+require 'timeout'
+require 'posix/spawn'
 
 module Guard
   class RackRunner
 
-    MAX_WAIT_COUNT = 10
-
-    attr_reader :options
+    attr_reader :options, :pid
 
     def initialize(options)
       @options = options
     end
 
-    def kill pid
-      system %{kill -INT #{pid}}
-      $?.exitstatus
+    def kill(pid, force = false)
+      result = -1
+
+      UI.debug("Trying to kill Rack (PID #{pid})...")
+      if !force
+        Process.kill("INT", pid)
+        begin
+          Timeout.timeout(options[:timeout]) do
+            _, status = Process.wait2(pid)
+            result = status.exitstatus
+            UI.debug("Killed Rack (Exit status: #{result})")
+          end
+        rescue Timeout::Error
+          UI.debug("Couldn't kill Rack with INT, switching to TERM")
+          force = true
+        end
+      end
+
+      if force
+        Process.kill("TERM", pid)
+      end
+
+      result
     end
 
     def start
       kill_unmanaged_pid! if options[:force_run]
-      run_rack_command!
-      wait_for_pid
+      @pid = run_rack_command!
+      true
     end
 
     def stop
-      # Rely on kill_unmanaged_pid if there's no pid file
-      return true unless File.file?(pid_file)
- 
-      if kill(pid) == 0
-        wait_for_no_pid
-        remove_pid_file
-      else
-        UI.info "Rackup exited with non-zero exit status whilst trying to stop."
+      # Rely on kill_unmanaged_pid if there's no pid
+      return true unless @pid
+
+      exitstatus = kill(@pid)
+      @pid = nil
+      if exitstatus != 0
+        UI.info "Rackup exited with non-zero exit status (#{exitstatus}) whilst trying to stop."
         return false
       end
 
@@ -42,55 +61,31 @@ module Guard
     end
 
     def build_rack_command
-      rack_options = [
-				options[:config],
-        '--env', options[:environment],
-        '--port', options[:port],
-        '--pid', pid_file
-      ]
+      command = %w{rackup}
+      command.push(
+        options[:config],
+        '--env', options[:environment].to_s,
+        '--port', options[:port].to_s
+      )
 
-      rack_options << '--daemonize' if options[:daemon]
-      rack_options << '--debug' if options[:debugger]
-      rack_options << ["--server",options[:server]] if options[:server]
+      command << '--daemonize' if options[:daemon]
+      command << '--debug' if options[:debugger]
+      command.push("--server", options[:server].to_s) if options[:server]
 
-      %{sh -c 'cd #{Dir.pwd} && rackup #{rack_options.join(' ')} &'}
-    end
-
-    def pid_file
-      File.expand_path(".guard-rack-#{options[:environment]}.pid")
-    end
-
-    def pid
-      File.file?(pid_file) ? File.read(pid_file).to_i : nil
-    end
-
-    def remove_pid_file
-      FileUtils.rm pid_file if File.exist? pid_file
-    end
-
-    def sleep_time
-      options[:timeout].to_f / MAX_WAIT_COUNT.to_f
+      command
     end
 
     private
-    
+
     def run_rack_command!
-      system build_rack_command
-    end
-
-    def has_pid?
-      File.file?(pid_file)
-    end
-
-    def wait_for_pid_action
-      sleep sleep_time
+      command = build_rack_command
+      UI.debug("Running Rack with command: #{command.inspect}")
+      POSIX::Spawn.spawn(*command)
     end
 
     def kill_unmanaged_pid!
       if pid = unmanaged_pid
-        kill pid
-        wait_for_no_pid
-        remove_pid_file
+        kill(pid, true)
       end
     end
 
@@ -102,25 +97,5 @@ module Guard
       }
       nil
     end
-
-    private
-
-      def wait_for_pid
-        wait_for_pid_loop
-      end
-
-      def wait_for_no_pid
-        wait_for_pid_loop(false)
-      end
-
-      def wait_for_pid_loop(check_for_existince = true)
-        count = 0
-        while !(check_for_existince ? has_pid? : !has_pid?) && count < MAX_WAIT_COUNT
-          wait_for_pid_action
-          count += 1
-        end
-        !(count == MAX_WAIT_COUNT)
-      end
-      
   end
 end