+++ spec/lib/guard/runner_spec.rb
@@ -1,32 +1,33 @@
 require 'spec_helper'
 require 'guard/rack/runner'
-require 'fakefs/spec_helpers'
 
 describe Guard::RackRunner do
   let(:runner) { Guard::RackRunner.new(options) }
   let(:environment) { 'development' }
   let(:port) { 3000 }
-  
+
   let(:default_options) { { :environment => environment, :port => port, :config => 'config.ru' } }
   let(:options) { default_options }
-  
-  describe '#pid' do
-    include FakeFS::SpecHelpers
 
-    context 'pid file exists' do
-      let(:pid) { 12345 }
+  before do
+    Guard::UI.stubs(:debug)
+  end
+
+  describe '#pid' do
+    context 'after running' do
+      let(:pid) { 1234 }
 
       before do
-        FileUtils.mkdir_p File.split(runner.pid_file).first
-        File.open(runner.pid_file, 'w') { |fh| fh.print pid }
+        POSIX::Spawn.stubs(:spawn).returns(pid)
+        runner.start
       end
 
-      it "should read the pid" do
+      it "should not be nil" do
         runner.pid.should == pid
       end
     end
 
-    context 'pid file does not exist' do
+    context 'before running' do
       it "should return nil" do
         runner.pid.should be_nil
       end
@@ -36,7 +37,7 @@ describe Guard::RackRunner do
   describe '#build_rack_command' do
     context 'no daemon' do
       it "should not have a daemon switch" do
-        runner.build_rack_command.should_not match(%r{ --daemonize})
+        runner.build_rack_command.should_not include('--daemonize')
       end
     end
 
@@ -44,54 +45,59 @@ describe Guard::RackRunner do
       let(:options) { default_options.merge(:daemon => true) }
 
       it "should have a daemon switch" do
-        runner.build_rack_command.should match(%r{ --daemonize})
+        runner.build_rack_command.should include('--daemonize')
       end
     end
-    
+
     context 'debugger' do
       let(:options) { default_options.merge(:debugger => true) }
 
       it "should have a debugger switch" do
-        runner.build_rack_command.should match(%r{ --debug})
+        runner.build_rack_command.should include('--debug')
       end
     end
+
     context "server" do
       let(:options) { default_options.merge(:server => "thin") }
 
       it "should honour server switch" do
-        runner.build_rack_command.should match(%r{ --server thin})
+        command = runner.build_rack_command
+        index = command.index('--server')
+        index.should be >= 0
+        command[index + 1].should == 'thin'
       end
     end
 
-		context "config file" do
-			context "default" do
-				it "should default to config.ru" do
-					runner.build_rack_command.should match(%r{ config.ru})
-				end
-			end
-			context "custom" do
-				let(:options) { default_options.merge(:config => 'config2.ru') }
-				it "should honour config option" do
-					options = default_options.merge(:config => 'config2.ru')
-					runner.build_rack_command.should match(%r{ config2.ru})
-				end
-			end
-		end
+    context "config file" do
+      context "default" do
+        it "should default to config.ru" do
+          runner.build_rack_command.should include('config.ru')
+        end
+      end
+
+      context "custom" do
+        let(:options) { default_options.merge(:config => 'config2.ru') }
+        it "should honour config option" do
+          options = default_options.merge(:config => 'config2.ru')
+          runner.build_rack_command.should include('config2.ru')
+        end
+      end
+    end
   end
 
   describe '#start' do
-    let(:kill_expectation) { runner.expects(:kill_unmanaged_pid!) }
-    let(:pid_stub) { runner.stubs(:has_pid?) }
+    let(:unmanaged_pid) { 4567 }
+    let(:pid) { 1234 }
+    let(:kill_expectation) { Process.expects(:kill).with("TERM", unmanaged_pid) }
 
     before do
-      runner.expects(:run_rack_command!).once
+      POSIX::Spawn.expects(:spawn).once.returns(pid)
+      runner.stubs(:unmanaged_pid).returns(unmanaged_pid)
     end
 
     context 'do not force run' do
       before do
-        pid_stub.returns(true)
         kill_expectation.never
-        runner.expects(:wait_for_pid_action).never
       end
 
       it "should act properly" do
@@ -103,47 +109,36 @@ describe Guard::RackRunner do
       let(:options) { default_options.merge(:force_run => true) }
 
       before do
-        pid_stub.returns(true)
         kill_expectation.once
-        runner.expects(:wait_for_pid_action).never
+        Process.expects(:wait2).never # don't wait on non-child processes
       end
 
       it "should act properly" do
         runner.start.should be_true
       end
     end
-
-    context "don't write the pid" do
-      before do
-        pid_stub.returns(false)
-        kill_expectation.never
-        runner.expects(:wait_for_pid_action).times(Guard::RackRunner::MAX_WAIT_COUNT)
-      end
-
-      it "should act properly" do
-        runner.start.should be_false
-      end
-    end
   end
 
   describe '#stop' do
-    
-    include FakeFS::SpecHelpers
 
-    context 'pid file exists' do
+    context 'pid exists' do
       let(:pid) { 12345 }
-      let(:kill_stub) { runner.stubs(:kill) }
-      let(:pid_stub) { runner.stubs(:has_pid?) }
+      let(:status_stub) { stub('process exit status') }
+      let(:wait_stub) { Process.stubs(:wait2) }
 
       before do
-        FileUtils.mkdir_p File.split(runner.pid_file).first
-        File.open(runner.pid_file, 'w') { |fh| fh.print pid }
+        POSIX::Spawn.stubs(:spawn).returns(pid)
+        runner.start
+
+        Process.expects(:kill).with('INT', pid)
       end
 
       context 'rackup returns successful exit status' do
         before do
-          kill_stub.returns(0)
+          wait_stub.returns([pid, status_stub])
+          status_stub.stubs(:exitstatus).returns(0)
         end
+
         it 'should return true' do
           runner.stop.should be_true
         end
@@ -151,32 +146,43 @@ describe Guard::RackRunner do
 
       context 'rackup returns unsuccessful exit status' do
         before do
-          Guard::UI.expects(:info).with(regexp_matches(/.+/))
-          kill_stub.returns(1)
+          Guard::UI.stubs(:info)
+          wait_stub.returns([pid, status_stub])
+          status_stub.stubs(:exitstatus).returns(1)
         end
+
         it 'should return false' do
           runner.stop.should be_false
         end
+
         it 'should send some kind of message to UI.info' do
+          Guard::UI.expects(:info).with(regexp_matches(/.+/))
+          runner.stop
+        end
+      end
+
+      context 'kill times out' do
+        before do
+          Guard::UI.stubs(:info)
+          wait_stub.raises(Timeout::Error)
+          Process.expects(:kill).with("TERM", pid)
+        end
+
+        it 'should return false' do
+          runner.stop.should be_false
+        end
+
+        it 'should send some kind of message to UI.info' do
+          Guard::UI.expects(:info).with(regexp_matches(/.+/))
           runner.stop
         end
       end
     end
 
-    context "pid file does not exist" do
-      it "should return true if there's no pid file" do
+    context "pid does not exist" do
+      it "should return true" do
         runner.stop.should be_true
       end
-    end
-
-  end
-
-  describe '#sleep_time' do
-    let(:timeout) { 30 }
-    let(:options) { default_options.merge(:timeout => timeout) }
-
-    it "should adjust the sleep time as necessary" do
-      runner.sleep_time.should == (timeout.to_f / Guard::RackRunner::MAX_WAIT_COUNT.to_f)
     end
   end
 end