+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+Next Release
+============
+* [#11](https://github.com/dblock/guard-rack/pull/11): Use posix-spawn instead of system() - [@viking](https://github.com/viking).
+
 1.2.2 (12/19/2012)
 ==================
 * [#10](https://github.com/dblock/guard-rack/pull/10): Added ability to specify rackup config - [@kgfullerton](https://github.com/kgfullerton).